+++ packages/dd-trace/test/appsec/telemetry/user.spec.js
@@ -1,0 +1,53 @@
+'use strict'
+
+const telemetryMetrics = require('../../../src/telemetry/metrics')
+const appsecNamespace = telemetryMetrics.manager.namespace('appsec')
+
+const appsecTelemetry = require('../../../src/appsec/telemetry')
+
+describe('Appsec User Telemetry metrics', () => {
+  let count, inc
+
+  beforeEach(() => {
+    inc = sinon.spy()
+    count = sinon.stub(appsecNamespace, 'count').returns({
+      inc
+    })
+
+    appsecNamespace.metrics.clear()
+    appsecNamespace.distributions.clear()
+  })
+
+  afterEach(sinon.restore)
+
+  describe('if enabled', () => {
+    beforeEach(() => {
+      appsecTelemetry.enable({
+        enabled: true,
+        metrics: true
+      })
+    })
+
+    describe('incrementMissingUserLoginMetric', () => {
+      it('should increment instrum.user_auth.missing_user_login metric', () => {
+        appsecTelemetry.incrementMissingUserLoginMetric('passport-local', 'login_success')
+
+        expect(count).to.have.been.calledOnceWithExactly('instrum.user_auth.missing_user_login', {
+          framework: 'passport-local',
+          event_type: 'login_success'
+        })
+      })
+    })
+
+    describe('incrementMissingUserIdMetric', () => {
+      it('should increment instrum.user_auth.missing_user_id metric', () => {
+        appsecTelemetry.incrementMissingUserIdMetric('passport', 'authenticated_request')
+
+        expect(count).to.have.been.calledOnceWithExactly('instrum.user_auth.missing_user_id', {
+          framework: 'passport',
+          event_type: 'authenticated_request'
+        })
+      })
+    })
+  })
+})
